requirements.txt
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cx_Freeze==7.2.5
cx_Logging==3.2.1;sys_platform=='win32'
lief==0.15.1
logger==1.4
loguru==0.7.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pyasn1==0.6.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pystray==0.19.5
pywin32==308;sys_platform=='win32'
rsa==4.9
six==1.16.0
win32-setctime==1.1.0;sys_platform=='win32'
h11==0.14.0
idna==3.10
lief==0.15.1
outcome==1.3.0.post0
pillow==11.0.0
pyasn1==0.6.1
pycparser==2.22
pyfiglet==1.0.2

PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
QDarkStyle==3.2.3
QtPy==2.4.2
requests==2.32.3
rsa==4.9
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
pillow_heif
whatimage
